some more work to use DParser accept @() user defined attributes better support template mixins support "alias this = identifier" and "alias ident = type" accept deprecated(expressions) install as extension: fix cache update

+++ vdc/ast/node.d
@@ -338,7 +338,7 @@ class Node
 		{
 			Node n = members[m];
 			narray[0] = n;
-			int mlen = members.length;
+			size_t mlen = members.length;
 			Node[] nm = n.expandNonScopeBlock(sc, narray);
 			assert(members.length == mlen);
 			if(nm.length == 1 && nm[0] == n)
